Web application initialization
Add: Web app project

Planning to separate the business/domain logic into separate solution.
It will act as an Web API for web and local application to use it's
components, the concrete applications will only use the effects and
maybe apply some specific logic to the results.
The main objective is to have common, but separated logic.